add packet handler
